[ANDROSDK-481] EntityDIModules
