+++ core/src/main/java/org/hisp/dhis/android/core/user/IsUserLoggedInCallable.java
