+++ core/src/main/java/org/hisp/dhis/android/core/user/UserCredentialsEntityDIModule.java
@@ -1,7 +1,7 @@
 /*
- * Copyright (c) 2004-2018, University of Oslo
- * All rights reserved.
+ * Copyright (c) 2017, University of Oslo
  *
+ * All rights reserved.
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  * Redistributions of source code must retain the above copyright notice, this
@@ -28,45 +28,27 @@
 
 package org.hisp.dhis.android.core.user;
 
-import org.hisp.dhis.android.core.calls.Call;
+import org.hisp.dhis.android.core.arch.handlers.IdentifiableSyncHandlerImpl;
+import org.hisp.dhis.android.core.arch.handlers.SyncHandler;
+import org.hisp.dhis.android.core.common.IdentifiableObjectStore;
 import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
-import org.hisp.dhis.android.core.wipe.WipeableModule;
 
-import javax.inject.Inject;
-import javax.inject.Provider;
-
+import dagger.Module;
+import dagger.Provides;
 import dagger.Reusable;
 
-@Reusable
-public final class UserInternalModule implements WipeableModule {
+@Module
+public final class UserCredentialsEntityDIModule {
 
-    private final DatabaseAdapter databaseAdapter;
-    public final UserModule publicModule;
-
-    // TODO delete after migrating MetadataCall to Dagger
-    public final Provider<Call<User>> userCallProvider;
-
-    @Inject
-    UserInternalModule(DatabaseAdapter databaseAdapter,
-                       UserModule publicModule,
-                       Provider<Call<User>> userCallProvider) {
-        this.databaseAdapter = databaseAdapter;
-        this.publicModule = publicModule;
-        this.userCallProvider = userCallProvider;
+    @Provides
+    @Reusable
+    IdentifiableObjectStore<UserCredentials> store(DatabaseAdapter databaseAdapter) {
+        return UserCredentialsStore.create(databaseAdapter);
     }
 
-    @Override
-    public void wipeMetadata() {
-        UserStore.create(databaseAdapter).delete();
-        UserCredentialsStore.create(databaseAdapter).delete();
-        UserOrganisationUnitLinkStore.create(databaseAdapter).delete();
-        AuthenticatedUserStore.create(databaseAdapter).delete();
-        AuthorityStore.create(databaseAdapter).delete();
-        new UserRoleStoreImpl(databaseAdapter).delete();
-    }
-
-    @Override
-    public void wipeData() {
-        // No data to wipe
+    @Provides
+    @Reusable
+    public static SyncHandler<UserCredentials> handler(IdentifiableObjectStore<UserCredentials> store) {
+        return new IdentifiableSyncHandlerImpl<>(store);
     }
 }
