+++ core/src/main/java/org/hisp/dhis/android/core/user/UserEntityDIModule.java
@@ -28,30 +28,20 @@
 
 package org.hisp.dhis.android.core.user;
 
-import org.hisp.dhis.android.core.arch.handlers.IdentifiableSyncHandlerImpl;
 import org.hisp.dhis.android.core.arch.handlers.SyncHandler;
-import org.hisp.dhis.android.core.calls.Call;
 import org.hisp.dhis.android.core.common.IdentifiableObjectStore;
-import org.hisp.dhis.android.core.common.ObjectWithoutUidStore;
 import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
 
 import dagger.Module;
 import dagger.Provides;
 import dagger.Reusable;
-import retrofit2.Retrofit;
 
 @Module
-public final class UserDIModule {
+public final class UserEntityDIModule {
 
     @Provides
     @Reusable
-    UserOrganisationUnitLinkStoreInterface userOrganisationUnitLinkStore(DatabaseAdapter databaseAdapter) {
-        return UserOrganisationUnitLinkStore.create(databaseAdapter);
-    }
-
-    @Provides
-    @Reusable
-    IdentifiableObjectStore<User> userStore(DatabaseAdapter databaseAdapter) {
+    IdentifiableObjectStore<User> store(DatabaseAdapter databaseAdapter) {
         return UserStore.create(databaseAdapter);
     }
 
@@ -59,35 +49,5 @@ public final class UserDIModule {
     @Reusable
     SyncHandler<User> handler(UserHandler userHandler) {
         return userHandler;
-    }
-
-    @Provides
-    @Reusable
-    Call<User> call(UserCall userCall) {
-        return userCall;
-    }
-
-    @Provides
-    @Reusable
-    UserService service(Retrofit retrofit) {
-        return retrofit.create(UserService.class);
-    }
-
-    @Provides
-    @Reusable
-    public static SyncHandler<UserCredentials> create(IdentifiableObjectStore<UserCredentials> store) {
-        return new IdentifiableSyncHandlerImpl<>(store);
-    }
-
-    @Provides
-    @Reusable
-    IdentifiableObjectStore<UserCredentials> userCredentialsStore(DatabaseAdapter databaseAdapter) {
-        return UserCredentialsStore.create(databaseAdapter);
-    }
-
-    @Provides
-    @Reusable
-    ObjectWithoutUidStore<AuthenticatedUserModel> authenticatedUserStore(DatabaseAdapter databaseAdapter) {
-        return AuthenticatedUserStore.create(databaseAdapter);
     }
 }
